Clean up tests

+++ core/cas-server-core-tickets/src/test/java/org/apereo/cas/ticket/registry/AbstractTicketRegistryTests.java
@@ -29,7 +29,9 @@ public abstract class AbstractTicketRegistryTests {
     @Before
     public void setUp() throws Exception {
         this.ticketRegistry = this.getNewTicketRegistry();
-        this.ticketRegistry.deleteAll();
+        if (ticketRegistry != null) {
+            this.ticketRegistry.deleteAll();
+        }
     }
 
     /**
